## Issue
Leadership transfer extension is an important feature in raft protocol. It is similar to leader abdication but it aims at switching to a pre-set node as fast as possible.  It can improve the availability of elasticsearch in the below situations:

1. In the case of a rolling restart, we shut down one node at a time. When we shut down the leader node, it takes at least 3 seconds or more downtime for the cluster to detect leader failure and elect a new leader. We can transfer the leader to a node that has been restarted before shut down the origin leader within a hundred milliseconds,  to avoid restarting the leader node. 

2. When the leader node is under heavy load but still connected with other nodes, this affects the health of the cluster. We need to switch the leader node to the new one as soon as possible instead of shutting it down and waiting for slow leader fault detection and re-election.

3. 2 master/data nodes plus one dedicated master node is a cheap and reasonable setup for many users.  The master/data nodes are only elected as leader temporarily when the dedicated master node fails down. In general situations, the low-spec dedicated master node is always elected as leader by leadership transfer extension.

## Solution
 The process of Leadership transfer extension is as follows.
1. Update cluster setting `preferred_master_name` to designate a follower as the future leader.
2. POST `_cluster/reelect` to the cluster.
3. The assigned node has the priority to become leader in the next round election. It begins to send votes to others to elect itself.
4. if the assigned node is elected as leader, the process ends. if the assigned node fails, it loses its priority, an equal election will be launched.


## Testing
Using Leadership transfer extension to avoid shutting down the active leader. Here is the test result.

|          | before optimization  | after optimization     | after optimization but the first election failed | 
|----------|----------|----------|----------| 
| downtime | 3.0-3.1s | 0.5-0.6s |1.0-1.1s    |