@boicehuang 2020-07-22
@ywelsch  Thanks for your reply. 
1. `cluster.fault_detection.leader_check.interval` defaults to 1s and `cluster.fault_detection.leader_check.retry_count`  defaults to 3. 
So 3 failed checks must happen before a leader is considered to have failed.
2.  In our cases,  Elasticsearch is used as a storage layer of the billing data. Since the timeout of writing our payment data to the storage is 1 second, reelection will have a great impact on our payment business.