@DaveCTurner 2020-07-22
As Yannick says, a clean shutdown doesn't need any timeouts, it causes an immediate reaction in Elasticsearch. In the case of an _unclean_ shutdown you should be relying on the OS to detect that the connection has dropped since the OS can react to this much more quickly than Elasticsearch. If you are not detecting unclean shutdowns fast enough then I suspect you have not configured `net.ipv4.tcp_retries2` appropriately; [it should be `3`](https://access.redhat.com/solutions/726753) if you want to detect a dropped connection in under a second. Relates https://github.com/elastic/elasticsearch/pull/59222.

The default for `cluster.fault_detection.leader_check.interval` is `10s` not `1s` and [you should not adjust this setting](https://www.elastic.co/guide/en/elasticsearch/reference/current/modules-discovery-settings.html#_expert_settings):

> If you adjust these settings then your cluster may not form correctly or may become unstable or intolerant of certain failures.

With such short timeouts your cluster is _very_ sensitive to transient disruptions and likely does more work reacting to those disruptions than it needs to.

Also, reiterating Yannick's comment above, it is very risky to set such a short indexing timeout. I don't think we should be considering adding complexity to Elasticsearch to address such unreasonably tight latency goals. This is certainly not the only reason why indexing might take longer than a second sometimes.