@DaveCTurner 2020-07-22
Apologies, I misread the setting name, `cluster.fault_detection.leader_check.timeout` defaults to `10s` but the `interval` is indeed `1s` by default. I'm confused, however, an unclean shutdown would require the leader checks to timeout (taking over 30 seconds by default) and a clean shutdown would trigger a new election immediately, so I do not understand what is taking 3 seconds.

Can you set `logger.org.elasticsearch.discovery: TRACE` and `logger.org.elasticsearch.cluster.coordination: TRACE` and share the resulting logs from one of these 3-second-long failovers?